Added CSS styling to login and create account windows
