+++ static/styles/styles.css
@@ -10,6 +10,7 @@ body {
     padding: 0;
     display: flex;
     flex-direction: column;
+    font-family: Arial, Helvetica, sans-serif;
 }
 
 .content {
@@ -82,21 +83,6 @@ body {
     z-index: 1;
 }
 
-form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 10px;
-    width: 50%;
-    height: 50%;
-    background-color: white;
-}
-
-.container {
-    padding: 10px;
-}
-
 .close-button {
     color:rgb(216, 215, 215);
     font-size: 15px;
@@ -106,6 +92,32 @@ form {
     position: absolute;
 }
 
+form {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    background: #FFFFFF;
+    width: 60%;
+    margin: 0 auto 100px;
+    padding: 45px;
+    text-align: center;
+    box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.2), 0 5px 5px 0 rgba(0, 0, 0, 0.24);
+}
+input[type=text] {
+    width: 100%;
+    padding: 12px 20px;
+    margin: 8px 0;
+    box-sizing: border-box;
+    font-size: 1.3em;
+}
+
+input[type=password] {
+    width: 100%;
+    padding: 12px 20px;
+    margin: 8px 0;
+    box-sizing: border-box;
+    font-size: 1.3em;
+}
